+++ contract-macro/tests/common/contract.rs
@@ -1,0 +1,18 @@
+use dusk_forge_contract::contract;
+
+#[contract]
+mod my_contract {
+    pub struct MyContract {
+        value: u64,
+    }
+
+    impl MyContract {
+        pub const fn new() -> Self {
+            Self { value: 0 }
+        }
+
+        pub fn get_value(&self) -> u64 {
+            self.value
+        }
+    }
+}